+++ examples/data-sources/morpheus_execute_schedule/data-source.tf
@@ -1,0 +1,3 @@
+data "morpheus_execute_schedule" "daily_at_midnight" {
+  name = "Run Daily at Midnight"
+}