+++ examples/resources/morpheus_execute_schedule/import.sh
@@ -1,0 +1,1 @@
+terraform import morpheus_execute_schedule.tf_example_execute_schedule 1